routes 90%

+++ app/products/repositories/product_category_repository.py
@@ -1,6 +1,7 @@
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 from app.products.models import ProductCategory
+from app.products.exceptions import ProductCategoryNotFoundException
 
 
 class ProductCategoryRepository:
@@ -17,10 +18,21 @@ class ProductCategoryRepository:
             return product_category
         except IntegrityError as e:
             raise e
+        except Exception as e:
+            raise e
 
     def read_by_id(self, product_category_id: str):
         product_category = self.db.query(ProductCategory).filter(ProductCategory.product_category_id ==
                                                                  product_category_id).first()
+        if product_category is None:
+            raise ProductCategoryNotFoundException\
+                (f"Product category with provided id: {product_category_id} not found.", 400)
+        return product_category
+
+    def read_by_name(self, name: str):
+        product_category = self.db.query(ProductCategory).filter(ProductCategory.name == name).first()
+        if product_category is None:
+            raise ProductCategoryNotFoundException(f"Product category with provided name: {name} not found.", 400)
         return product_category
 
     def read_all(self):
@@ -31,6 +43,9 @@ class ProductCategoryRepository:
         try:
             product_category = self.db.query(ProductCategory).filter(ProductCategory.product_category_id ==
                                                                      product_category_id).first()
+            if product_category is None:
+                raise ProductCategoryNotFoundException \
+                    (f"Product category with provided id: {product_category_id} not found.", 400)
             self.db.delete(product_category)
             self.db.commit()
             return True
@@ -41,6 +56,9 @@ class ProductCategoryRepository:
         try:
             product_category = self.db.query(ProductCategory).filter(ProductCategory.product_category_id ==
                                                                      product_category_id).first()
+            if product_category is None:
+                raise ProductCategoryNotFoundException \
+                    (f"Product category with provided id: {product_category_id} not found.", 400)
             product_category.name = new_name
             self.db.add(product_category)
             self.db.commit()
@@ -48,7 +66,3 @@ class ProductCategoryRepository:
             return product_category
         except Exception as e:
             raise e
-
-    def read_by_name(self, name: str):
-        product_category = self.db.query(ProductCategory).filter(ProductCategory.name == name).first()
-        return product_category
