+++ app/users/repositories/user_repository.py
@@ -1,6 +1,7 @@
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 from app.users.models import User
+from app.users.exceptions import UserNotFoundException
 
 
 class UserRepository:
@@ -17,6 +18,8 @@ class UserRepository:
             return user
         except IntegrityError as e:
             raise e
+        except Exception as e:
+            raise e
 
     def create_superuser(self, email, password):
         try:
@@ -27,9 +30,19 @@ class UserRepository:
             return user
         except IntegrityError as e:
             raise e
+        except Exception as e:
+            raise e
 
     def read_by_id(self, user_id: str):
         user = self.db.query(User).filter(User.user_id == user_id).first()
+        if user is None:
+            raise UserNotFoundException(f"User with provided id: {user_id} not found.", 400)
+        return user
+
+    def read_by_email(self, email: str):
+        user = self.db.query(User).filter(User.email == email).first()
+        if user is None:
+            raise UserNotFoundException(f"User with provided email: {email} not found.", 400)
         return user
 
     def read_all(self):
@@ -39,6 +52,8 @@ class UserRepository:
     def delete_by_id(self, user_id: str):
         try:
             user = self.db.query(User).filter(User.user_id == user_id).first()
+            if user is None:
+                raise UserNotFoundException(f"User with provided id: {user_id} not found.", 400)
             self.db.delete(user)
             self.db.commit()
             return True
@@ -48,6 +63,8 @@ class UserRepository:
     def update_active(self, user_id: str, active: bool):
         try:
             user = self.db.query(User).filter(User.user_id == user_id).first()
+            if user is None:
+                raise UserNotFoundException(f"User with provided id: {user_id} not found.", 400)
             user.active = active
             self.db.add(user)
             self.db.commit()
@@ -55,7 +72,3 @@ class UserRepository:
             return user
         except Exception as e:
             raise e
-
-    def read_by_email(self, email: str):
-        user = self.db.query(User).filter(User.email == email).first()
-        return user
